refactor: converge sdk transport assembly and service contracts (#132)
