+++ tests/test_settings.py
@@ -40,7 +40,6 @@ def test_settings_parse_ops_flags_and_timeouts():
         "A2A_ENABLE_HEALTH_ENDPOINT": "false",
         "A2A_ENABLE_SESSION_SHELL": "false",
         "A2A_CANCEL_ABORT_TIMEOUT_SECONDS": "0.25",
-        "A2A_STREAM_SSE_PING_SECONDS": "8",
         "A2A_STREAM_IDLE_DIAGNOSTIC_SECONDS": "45",
         "A2A_INTERRUPT_REQUEST_TTL_SECONDS": "90",
     }
@@ -49,7 +48,6 @@ def test_settings_parse_ops_flags_and_timeouts():
         assert settings.a2a_enable_health_endpoint is False
         assert settings.a2a_enable_session_shell is False
         assert settings.a2a_cancel_abort_timeout_seconds == 0.25
-        assert settings.a2a_stream_sse_ping_seconds == 8
         assert settings.a2a_stream_idle_diagnostic_seconds == 45
         assert settings.a2a_interrupt_request_ttl_seconds == 90
 
@@ -101,28 +99,6 @@ def test_settings_reject_invalid_interrupt_request_ttl():
         with pytest.raises(ValidationError) as excinfo:
             Settings()
     assert "A2A_INTERRUPT_REQUEST_TTL_SECONDS" in str(excinfo.value)
-
-
-def test_settings_reject_invalid_stream_ping_seconds():
-    env = {
-        "A2A_BEARER_TOKEN": "test",
-        "A2A_STREAM_SSE_PING_SECONDS": "0",
-    }
-    with mock.patch.dict(os.environ, env, clear=True):
-        with pytest.raises(ValidationError) as excinfo:
-            Settings()
-    assert "A2A_STREAM_SSE_PING_SECONDS" in str(excinfo.value)
-
-
-def test_settings_reject_non_integer_stream_ping_seconds():
-    env = {
-        "A2A_BEARER_TOKEN": "test",
-        "A2A_STREAM_SSE_PING_SECONDS": "8.5",
-    }
-    with mock.patch.dict(os.environ, env, clear=True):
-        with pytest.raises(ValidationError) as excinfo:
-            Settings()
-    assert "A2A_STREAM_SSE_PING_SECONDS" in str(excinfo.value)
 
 
 def test_settings_reject_invalid_stream_idle_diagnostic_seconds():
